update. #20 Allow builds to be done without cgo (pkcs11 will no longer be available).
