+++ auth.go
@@ -18,7 +18,6 @@ import (
 	"strings"
 
 	"github.com/ScaleFT/sshkeys"
-	"github.com/miekg/pkcs11/p11"
 	"golang.org/x/crypto/ssh"
 	"golang.org/x/crypto/ssh/agent"
 )
@@ -161,80 +160,6 @@ func CreateSignerCertificate(cert string, keySigner ssh.Signer) (certSigner ssh.
 	certSigner, err = ssh.NewCertSigner(certificate, keySigner)
 	if err != nil {
 		return
-	}
-
-	return
-}
-
-// CreateAuthMethodPKCS11 return []ssh.AuthMethod generated from pkcs11 token.
-// PIN is required to generate a AuthMethod from a PKCS 11 token.
-//
-// WORNING: Does not work if multiple tokens are stuck at the same time.
-func CreateAuthMethodPKCS11(provider, pin string) (auth []ssh.AuthMethod, err error) {
-	signers, err := CreateSignerPKCS11(provider, pin)
-	if err != nil {
-		return
-	}
-
-	for _, signer := range signers {
-		auth = append(auth, ssh.PublicKeys(signer))
-	}
-	return
-}
-
-// CreateSignerPKCS11 returns []ssh.Signer generated from PKCS11 token.
-// PIN is required to generate a Signer from a PKCS 11 token.
-//
-// WORNING: Does not work if multiple tokens are stuck at the same time.
-func CreateSignerPKCS11(provider, pin string) (signers []ssh.Signer, err error) {
-	// get absolute path
-	provider = getAbsPath(provider)
-
-	// Create p11.module
-	module, err := p11.OpenModule(provider)
-	if err != nil {
-		return
-	}
-
-	// Get p11 Module's Slot
-	slots, err := module.Slots()
-	if err != nil {
-		return
-	}
-	c11array := []*C11{}
-
-	for _, slot := range slots {
-		tokenInfo, err := slot.TokenInfo()
-		if err != nil {
-			continue
-		}
-
-		c := &C11{
-			Label: tokenInfo.Label,
-			PIN:   pin,
-		}
-		c11array = append(c11array, c)
-	}
-
-	// Destroy Module
-	module.Destroy()
-
-	// for loop
-	for _, c11 := range c11array {
-		err := c11.CreateCtx(provider)
-		if err != nil {
-			continue
-		}
-
-		sigs, err := c11.GetSigner()
-		if err != nil {
-			continue
-		}
-
-		for _, sig := range sigs {
-			signer, _ := ssh.NewSignerFromSigner(sig)
-			signers = append(signers, signer)
-		}
 	}
 
 	return
