+++ pkcs11.go
@@ -1,6 +1,8 @@
 // Copyright (c) 2020 Blacknon. All rights reserved.
 // Use of this source code is governed by an MIT license
 // that can be found in the LICENSE file.
+//go:build cgo
+// +build cgo
 
 package sshlib
 
@@ -9,6 +11,7 @@ import (
 )
 
 // C11 struct for Crypto11 processing.
+// Not available if cgo is disabled.
 type C11 struct {
 	Label string
 	PIN   string
@@ -16,6 +19,7 @@ type C11 struct {
 }
 
 // getPIN is set token's PIN Code to c.PIN
+// Not available if cgo is disabled.
 func (c *C11) getPIN() (err error) {
 	if c.PIN == "" {
 		c.PIN, err = getPassphrase(c.Label + "'s PIN:")
@@ -25,6 +29,7 @@ func (c *C11) getPIN() (err error) {
 }
 
 // CreateCtx is create crypto11.Context
+// Not available if cgo is disabled.
 func (c *C11) CreateCtx(provider string) (err error) {
 	// Get PIN Code
 	err = c.getPIN()
@@ -47,7 +52,8 @@ func (c *C11) CreateCtx(provider string) (err error) {
 	return
 }
 
-// GetSigner return []crypto11.Signer
+// GetSigner return []crypto11.Signer.
+// Not available if cgo is disabled.
 func (c *C11) GetSigner() (signer []crypto11.Signer, err error) {
 	return c.Ctx.FindAllKeyPairs()
 }
